src/analysis/match_predictor.py
import pandas as pd
from sklearn.neighbors import KNeighborsRegressor

class MatchPredictor:
    def __init__(self, historical_data_path):
        self.data = pd.read_csv(historical_data_path)
        self.model = KNeighborsRegressor(n_neighbors=5)
        
    def check_direct_matchup(self, team1, team2):
        direct_matches = self.data[
            ((self.data['home_team'] == team1) & (self.data['away_team'] == team2)) |
            ((self.data['home_team'] == team2) & (self.data['away_team'] == team1))
        ]
        return len(direct_matches) > 0
    
    def predict_match(self, home_team, away_team):
        has_direct_matchup = self.check_direct_matchup(home_team, away_team)
        
        if not has_direct_matchup:
            print(f"\nWarning: No direct matches found between {home_team} and {away_team}")
            print("Using similar team performance patterns for prediction...")
            return self.predict_from_similar_teams(home_team, away_team)
        
        return self.predict_from_direct_matches(home_team, away_team)
    
    def predict_from_direct_matches(self, home_team, away_team):
        # Add prediction logic for direct matches
        pass
    
    def predict_from_similar_teams(self, home_team, away_team):
        # Add prediction logic for similar teams
        pass