requirements.txt
# requirements.txt
pyspark==3.2.0
pandas==1.4.0
scikit-learn==1.0.2
numpy==1.21.0
matplotlib==3.5.0
seaborn==0.11.2
python-dotenv
beautifulsoup4
requests
kaggle
scikit-learn
jupyter
pyyaml==6.0
pytest==7.0.0
